refactor: registration  command/read 구조 통합
