+++ backend/src/main/java/com/rungo/api/domain/registration/service/RegistrationService.java
@@ -3,11 +3,14 @@ package com.rungo.api.domain.registration.service;
 import com.rungo.api.domain.marathon.course.entity.Course;
 import com.rungo.api.domain.marathon.course.repository.CourseRepository;
 import com.rungo.api.domain.marathon.marathon.entity.Marathon;
+import com.rungo.api.domain.marathon.marathon.repository.MarathonRepository;
 import com.rungo.api.domain.notification.event.RegistrationCompletedEvent;
 import com.rungo.api.domain.registration.dto.CreateRegistrationReq;
 import com.rungo.api.domain.registration.dto.CreateRegistrationRes;
+import com.rungo.api.domain.registration.dto.MyRegistrationRes;
 import com.rungo.api.domain.registration.entity.Registration;
 import com.rungo.api.domain.registration.entity.RegistrationCancelHistory;
+import com.rungo.api.domain.registration.enumtype.MyRegistrationStatusFilter;
 import com.rungo.api.domain.registration.repository.RegistrationCancelHistoryRepository;
 import com.rungo.api.domain.registration.repository.RegistrationRepository;
 import com.rungo.api.domain.users.entity.Users;
@@ -16,23 +19,37 @@ import com.rungo.api.global.exception.CustomException;
 import com.rungo.api.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class RegistrationCommandService {
+public class RegistrationService {
 
     private final RegistrationRepository registrationRepository;
     private final RegistrationCancelHistoryRepository registrationCancelHistoryRepository;
     private final CourseRepository courseRepository;
+    private final MarathonRepository marathonRepository;
     private final UserRepository userRepository;
     private final ApplicationEventPublisher eventPublisher;
 
     public CreateRegistrationRes create(Long userId, CreateRegistrationReq request) {
+
+        // 필수 약관 미동의
+        if (!request.agreedTerms()) {
+            throw new CustomException(ErrorCode.REGISTRATION_TERMS_REQUIRED);
+        }
 
         Users user = userRepository.findById(userId)
                 .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
@@ -41,10 +58,6 @@ public class RegistrationCommandService {
         Marathon marathon = course.getMarathon();
         LocalDateTime now = LocalDateTime.now();
 
-        // 필수 약관 미동의
-        if (!request.agreedTerms()) {
-            throw new CustomException(ErrorCode.REGISTRATION_TERMS_REQUIRED);
-        }
         // 접수 기간이 아니면 생성할 수 없다.
         if (now.isBefore(marathon.getRegistrationStartAt()) || now.isAfter(marathon.getRegistrationEndAt())) {
             throw new CustomException(ErrorCode.REGISTRATION_PERIOD_INVALID);
@@ -114,5 +127,63 @@ public class RegistrationCommandService {
         // 동시성 제어 미적용 메서드
         // registration.getCourse().decreaseCurrentCount();
         registrationRepository.delete(registration);
+    }
+
+    // status 필터에 따른 내 접수 목록 조회
+    // ACTIVE   : 정상 접수인 상태 (취소 되지 않은 모든 접수)
+    // CANCELED : 취소된 접수 상태
+    @Transactional(readOnly = true)
+    public MyRegistrationRes getMyRegistrations(Long userId, MyRegistrationStatusFilter status, Pageable pageable) {
+        if (status == MyRegistrationStatusFilter.CANCELED) {
+            return getCanceledRegistrations(userId, pageable);
+        }
+
+        return getActiveRegistrations(userId, pageable);
+    }
+
+    private MyRegistrationRes getActiveRegistrations(Long userId, Pageable pageable) {
+        Pageable sortedPageable = PageRequest.of(
+                pageable.getPageNumber(),
+                pageable.getPageSize(),
+                Sort.by(
+                        Sort.Order.desc("appliedAt"),
+                        Sort.Order.desc("id")
+                )
+        );
+
+        Page<Registration> page = registrationRepository.findByUser_Id(userId, sortedPageable);
+
+        return MyRegistrationRes.fromActive(page);
+    }
+
+    private MyRegistrationRes getCanceledRegistrations(Long userId, Pageable pageable) {
+        Pageable sortedPageable = PageRequest.of(
+                pageable.getPageNumber(),
+                pageable.getPageSize(),
+                Sort.by(
+                        Sort.Order.desc("canceledAt"),
+                        Sort.Order.desc("id")
+                )
+        );
+
+        Page<RegistrationCancelHistory> page = registrationCancelHistoryRepository.findByUserId(userId, sortedPageable);
+
+        List<Long> marathonIds = page.getContent().stream()
+                .map(RegistrationCancelHistory::getMarathonId)
+                .distinct()
+                .toList();
+
+        List<Long> courseIds = page.getContent().stream()
+                .map(RegistrationCancelHistory::getCourseId)
+                .distinct()
+                .toList();
+
+        Map<Long, Marathon> marathonMap = marathonRepository.findAllById(marathonIds).stream()
+                .collect(Collectors.toMap(Marathon::getId, Function.identity()));
+
+        Map<Long, Course> courseMap = courseRepository.findAllById(courseIds).stream()
+                .collect(Collectors.toMap(Course::getId, Function.identity()));
+
+        return MyRegistrationRes.fromCanceled(page, marathonMap, courseMap);
     }
 }
